backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
